+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.9.0] - 2025-10-23
+- Add an updater script that updates pip requirements, performs migrations, and collects static files.
+- Clear cache on app start.
+- Add new setting for diarization checkpoint path.
+- Track Python version in .python-version.
+- Print exceptions on process submission.
+- Update Django, pyannote.audio, and yt-dlp.
+- Update diarization code for pyannote.audio v4.0.0.
+- Disable diarize radio if settings are not configured.
+- Use Django timezone utility for date/time now.
+
 ## [1.8.0] - 2025-09-30
 - Disable media controls on edit page if no media exists.
 - Add button to cancel submissions. Hide button if transcription cannot be cancelled.